% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BS-class.R
\docType{class}
\name{BS-class}
\alias{BS-class}
\title{Black & Scholes}
\description{
An S4 class to represent a Black & Scholes assets built upon G2 class
}
\section{Slots}{

\describe{
\item{\code{g2model}}{G2. Trained and projected G2 model}

\item{\code{s0}}{Numeric. Initial value of the asset.}

\item{\code{vol}}{Numeric. Periodic volatility.}

\item{\code{div}}{Numeric. Percentage of divividend each year.}

\item{\code{rho}}{Numeric. Correlation with the rate model.}

\item{\code{s}}{Matrix. Projection of the asset.}
}}

