% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G2-class.R
\docType{methods}
\name{calibrate}
\alias{calibrate}
\title{Rate model calibration}
\usage{
calibrate(.Object, swaptions, maxIter = 1000, monitor = TRUE)
}
\description{
\code{calibrate} calibrates the model on swaptions prices.
}
\details{
@param swaptions Swaptions. List of swaptions under the format of the class Swaption.
 @param maxiter Numeric. Max number of iterations for Nelder-Mead algorithm. Default to 1000.
 @param monitor Boolean. To display intermediary results. Default to TRUE.

@return The calibrated model
}
