% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VSK-class.R
\name{draw}
\alias{draw}
\title{Draw method for VSK}
\usage{
draw(x, type = "index")
}
\arguments{
\item{x}{the VSK object to plot}

\item{type}{Type of plot. "index" or "rate"}
}
\description{
Plot the trajectories of the model.
}
