% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G2-class.R
\docType{class}
\name{G2-class}
\alias{G2-class}
\title{G2 Model}
\description{
An S4 class to represent a G2 ++ model.
}
\section{Slots}{

\describe{
\item{\code{curve}}{ZCCurve. Zero coupon curve as basis for projection.}

\item{\code{a}}{Numeric. G2++ parameter.}

\item{\code{b}}{Numeric. G2++ parameter.}

\item{\code{sigma}}{Numeric. G2++ parameter.}

\item{\code{eta}}{Numeric. G2++ parameter.}

\item{\code{rho}}{Numeric. G2++ parameter.}

\item{\code{horizon}}{Numeric. Projection horizon.}

\item{\code{nsimul}}{Numeric. Number of simulations.}

\item{\code{x}}{Matrix Projection of the first component.}

\item{\code{y}}{Matrix. Projection of the second component.}

\item{\code{r}}{Matrix. Short rate projection.}

\item{\code{Wx}}{Matrix. Generated normal distribution for x projection.}
}}

