% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rate2zcp}
\alias{rate2zcp}
\title{Zero Coupon Price}
\usage{
rate2zcp(mat, rate, method = "continuous")
}
\arguments{
\item{mat}{Vector. Maturities of the bound.}

\item{rate}{Vector. Panel of rates for which the price is to be computed.}

\item{method}{Character. Computation method:
\itemize{
  \item{"continuous": continuous rate. Default value.}
  \item{"actuarial": actuarial rate.}
  \item{"libor": libor rate.}}}
}
\value{
The zero coupon prices associated to the maturity.
}
\description{
\code{rate2zcp} computes the price of a zero coupon bond for given maturities and rates.
}
\examples{
rate2zcp(1, 0.01, "continuous")

}
