% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Swaptions-class.R
\docType{class}
\name{Swaptions-class}
\alias{Swaptions-class}
\title{Swaptions}
\description{
An S4 class to represent ATM swaptions for calibration of a G2++ model.
}
\section{Slots}{

\describe{
\item{\code{type}}{Character. Type of volatility:
\itemize{
  \item{"normal".}
  \item{"lognormal". Default value}}}

\item{\code{curve}}{ZCCurve. Zero coupon curve to use for pricing.}

\item{\code{mat}}{Vector. List of maturities.}

\item{\code{tenor}}{Vector. List of tenors.}

\item{\code{vol}}{Vector. List of volatilities corresponding to maturities and tenors.}

\item{\code{freq}}{Numeric. Number of payments by period.}

\item{\code{price}}{Vector. Swaptions prices.}
}}

