% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VSK-class.R
\name{trajv}
\alias{trajv}
\title{VSK: get trajectories}
\usage{
trajv(.Object, type = "index")
}
\arguments{
\item{.Object}{VSK Object whose trajectories are wanted.}

\item{type}{character. Type of trajectories: "index" or "rate"}
}
\value{
The trajectories as a matrix.
}
\description{
\code{trajv} returns the trajectories for the Vasicek model.
}
\examples{
index <- c(100,102.32,104.47,106.23,108.21,110.36,
111.66,112.4,112.97,114.84,116.7,119.03,121.46,124.01,126.33,
128.37,130.29,133.98,134.09,136.13,139.01,141.73,142.92,
143.66,143.71,144,145.41)
inflation <- vsk(index_histo = index)
inf_traj <- trajv(inflation)

}
