% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VSK-class.R
\docType{class}
\name{VSK-class}
\alias{VSK-class}
\alias{print,VSK-method}
\alias{show,VSK-method}
\title{Vasicek model}
\usage{
\S4method{print}{VSK}(x)

\S4method{show}{VSK}(object)
}
\arguments{
\item{x}{the VSK object to print}

\item{object}{the VSK object to show}
}
\description{
An S4 class to represent a value modelled by a Vasicek projection model. In the context of this esg, this model is thought for inflation modelling.
}
\section{Methods (by generic)}{
\itemize{
\item \code{print}: print method for VSK

\item \code{show}: show method for VSK
}}

\section{Slots}{

\describe{
\item{\code{q0}}{Numeric. Initial value of rate.}

\item{\code{a}}{Numeric. Speed of reversion.}

\item{\code{b}}{Numeric. Long-term mean for the model.}

\item{\code{sigma}}{Numeric. Periodic volatility.}

\item{\code{q}}{Matrix. Projection of the asset.}
}}

