% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VSK-class.R
\name{vsk}
\alias{vsk}
\title{VSK: constructor}
\usage{
vsk(q0 = NULL, a = NULL, b = NULL, horizon = 50, nsimul = 1000,
  sigma = NULL, index_histo = NULL, q_histo = NULL, W = NULL)
}
\arguments{
\item{q0}{Numeric. Initial value of rate. Optional if index_histo or q_histo is given.}

\item{a}{Numeric. Speed of reversion. Optional if index_histo or q_histo is given.}

\item{b}{Numeric. Long-term mean for the model. Optional if index_histo or q_histo is given.}

\item{horizon}{Numeric. Projection horizon. Default to 50.}

\item{nsimul}{Numeric. Number of simulations. Default to 1000.}

\item{sigma}{Numeric. Periodic volatility. Optional if index_histo or q_histo is given.}

\item{index_histo}{Vector. History of the index to modelize. Optional if q0, a, b, sigma or q_histo is given. All values from the intervall must be given. Any missing value should be set to NA.}

\item{q_histo}{Vector. History of rates to modelize. Optional if q0, a, b, sigma or index_histo is given. All values from the intervall must be given. Any missing value should be set to NA.}

\item{W}{Matrix. Projection of the asset via a normal distribution. Optional.}
}
\value{
The model as a class.
}
\description{
\code{vsk} creates a vasicek model object.
}
\details{
The rate at time t is computed as log(Index_t/Index_t-1)
}
\examples{
index <- c(100,102.32,104.47,106.23,108.21,110.36,
111.66,112.4,112.97,114.84,116.7,119.03,121.46,124.01,126.33,
128.37,130.29,133.98,134.09,136.13,139.01,141.73,142.92,
143.66,143.71,144,145.41)
inflation <- vsk(index_histo = index)

}
