% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZCCurve-class.R
\docType{class}
\name{ZCCurve-class}
\alias{ZCCurve-class}
\title{Zero Coupon Curve}
\description{
An S4 class to represent a zero coupon curve.
}
\section{Slots}{

\describe{
\item{\code{method}}{Character. Method for discounting the curve:
\itemize{
  \item{"continuous": continuous rate. Default value.}
  \item{"actuarial": actuarial rate}
  \item{"libor": libor rate}}}

\item{\code{rates}}{Vector. List of rates from maturity 1 to horizon.}

\item{\code{zcp}}{Vector. List of corresponding zero coupon bonds prices.}

\item{\code{ifr}}{Vector. Instant forward rates at time 0.}
}}

