% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{zcp2rate}
\alias{zcp2rate}
\title{Rate computation}
\usage{
zcp2rate(mat, zcp, method = "continuous")
}
\arguments{
\item{mat}{Vector. Maturities of the bound.}

\item{zcp}{Vector. Panel of zero coupon bonds prices for which the rates are to be computed.}

\item{method}{Character. Computation method:
\itemize{
  \item{"continuous": continuous rate. Default value.}
  \item{"actuarial": actuarial rate.}
  \item{"libor": libor rate.}}}
}
\value{
The rates associated to the maturities.
}
\description{
\code{zcp2rate} computes the rate equivalent to a zero coupon price for a maturity.
}
\examples{
zcp2rate(1, 0.9900498, "continuous")

}
