% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BS-class.R
\docType{class}
\name{BS-class}
\alias{BS-class}
\alias{plot,BS,missing-method}
\alias{print,BS-method}
\alias{show,BS-method}
\title{Black & Scholes}
\usage{
\S4method{plot}{BS,missing}(x, y)

\S4method{print}{BS}(x)

\S4method{show}{BS}(object)
}
\arguments{
\item{x}{the BS object to plot}

\item{y}{classical plot parameters. Useless here.}

\item{object}{the BS object to show}
}
\description{
An S4 class to represent a Black & Scholes assets built upon G2 class
}
\section{Methods (by generic)}{
\itemize{
\item \code{plot}: plot method for BS

\item \code{print}: print method for BS

\item \code{show}: show method for BS
}}

\section{Slots}{

\describe{
\item{\code{g2model}}{G2. Trained and projected G2 model}

\item{\code{s0}}{Numeric. Initial value of the asset.}

\item{\code{vol}}{Numeric. Periodic volatility.}

\item{\code{div}}{Numeric. Percentage of divividend each year.}

\item{\code{rho}}{Numeric. Correlation with the rate model.}

\item{\code{s}}{Matrix. Projection of the asset.}
}}

