% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G2-class.R
\name{calibrate}
\alias{calibrate}
\title{Rate model calibration}
\usage{
calibrate(.Object, swaptions, maxIter = 1000, monitor = TRUE)
}
\arguments{
\item{.Object}{G2 object. The object whose model is to be projected.}

\item{swaptions}{Swaptions. List of swaptions under the format of the class Swaption.}

\item{maxIter}{Numeric. Max number of iterations for Nelder-Mead algorithm. Default to 1000.}

\item{monitor}{Boolean. To display intermediary results. Default to TRUE.}
}
\value{
The calibrated model.
}
\description{
\code{calibrate} calibrates the model on swaptions prices.
}
