% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G2-class.R
\docType{class}
\name{G2-class}
\alias{G2-class}
\alias{plot,G2,missing-method}
\alias{print,G2-method}
\alias{show,G2-method}
\title{G2 Model}
\usage{
\S4method{plot}{G2,missing}(x, y)

\S4method{print}{G2}(x)

\S4method{show}{G2}(object)
}
\arguments{
\item{x}{the G2 object to plot}

\item{y}{classical plot parameters. Useless here.}

\item{object}{the G2 object to show}
}
\description{
An S4 class to represent a G2 ++ model.
}
\section{Methods (by generic)}{
\itemize{
\item \code{plot}: plot method for G2

\item \code{print}: print method for G2

\item \code{show}: show method for G2
}}

\section{Slots}{

\describe{
\item{\code{curve}}{ZCCurve. Zero coupon curve as basis for projection.}

\item{\code{a}}{Numeric. G2++ parameter.}

\item{\code{b}}{Numeric. G2++ parameter.}

\item{\code{sigma}}{Numeric. G2++ parameter.}

\item{\code{eta}}{Numeric. G2++ parameter.}

\item{\code{rho}}{Numeric. G2++ parameter.}

\item{\code{horizon}}{Numeric. Projection horizon.}

\item{\code{nsimul}}{Numeric. Number of simulations.}

\item{\code{x}}{Matrix Projection of the first component.}

\item{\code{y}}{Matrix. Projection of the second component.}

\item{\code{r}}{Matrix. Short rate projection.}

\item{\code{Wx}}{Matrix. Generated normal distribution for x projection.}
}}

