% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Swaptions-class.R
\docType{class}
\name{Swaptions-class}
\alias{Swaptions-class}
\alias{plot,Swaptions,missing-method}
\alias{print,Swaptions-method}
\alias{show,Swaptions-method}
\title{Swaptions}
\usage{
\S4method{plot}{Swaptions,missing}(x, y)

\S4method{print}{Swaptions}(x)

\S4method{show}{Swaptions}(object)
}
\arguments{
\item{x}{the Swaptions object to plot}

\item{y}{classical plot parameters. Useless here.}

\item{object}{the Swaptions object to show}
}
\description{
An S4 class to represent ATM swaptions for calibration of a G2++ model.
}
\section{Methods (by generic)}{
\itemize{
\item \code{plot}: plot method for Swaptions

\item \code{print}: print method for Swaptions

\item \code{show}: show method for Swaptions
}}

\section{Slots}{

\describe{
\item{\code{type}}{Character. Type of volatility:
\itemize{
  \item{"normal".}
  \item{"lognormal". Default value}}}

\item{\code{curve}}{ZCCurve. Zero coupon curve to use for pricing.}

\item{\code{mat}}{Vector. List of maturities.}

\item{\code{tenor}}{Vector. List of tenors.}

\item{\code{vol}}{Vector. List of volatilities corresponding to maturities and tenors.}

\item{\code{freq}}{Numeric. Number of payments by period.}

\item{\code{price}}{Vector. Swaptions prices.}
}}

